+++ src/Command/AbstractCommand.php
@@ -120,6 +120,7 @@ abstract class AbstractCommand implements CommandInterface
      */
     protected string|null $refreshTableName = null;
     protected Closure|null $retryHandler = null;
+    protected bool $dbTypecasting = true;
     protected bool $phpTypecasting = false;
     /**
      * @var string The SQL statement to execute.
@@ -457,13 +458,6 @@ abstract class AbstractCommand implements CommandInterface
         return is_scalar($result) ? $result : null;
     }
 
-    public function withPhpTypecasting(bool $phpTypecasting = true): static
-    {
-        $new = clone $this;
-        $new->phpTypecasting = $phpTypecasting;
-        return $new;
-    }
-
     public function renameColumn(string $table, string $oldName, string $newName): static
     {
         $sql = $this->getQueryBuilder()->renameColumn($table, $oldName, $newName);
@@ -527,6 +521,28 @@ abstract class AbstractCommand implements CommandInterface
     ): static {
         $sql = $this->getQueryBuilder()->upsert($table, $insertColumns, $updateColumns, $params);
         return $this->setSql($sql)->bindValues($params);
+    }
+
+    public function withDbTypecasting(bool $dbTypecasting = true): static
+    {
+        $new = clone $this;
+        $new->dbTypecasting = $dbTypecasting;
+        return $new;
+    }
+
+    public function withPhpTypecasting(bool $phpTypecasting = true): static
+    {
+        $new = clone $this;
+        $new->phpTypecasting = $phpTypecasting;
+        return $new;
+    }
+
+    public function withTypecasting(bool $typecasting = true): static
+    {
+        $new = clone $this;
+        $new->dbTypecasting = $typecasting;
+        $new->phpTypecasting = $typecasting;
+        return $new;
     }
 
     /**