+++ src/Command/CommandInterface.php
@@ -620,11 +620,6 @@ interface CommandInterface
     public function insertWithReturningPks(string $table, array $columns): array|false;
 
     /**
-     * Returns a copy of the instance with enabled or disabled typecasting of values when retrieving records from DB.
-     */
-    public function withPhpTypecasting(bool $phpTypecasting = true): static;
-
-    /**
      * Prepares the SQL statement to be executed.
      *
      * For complex SQL statement that's to be executed many times, this may improve performance.
@@ -894,4 +889,20 @@ interface CommandInterface
         bool|array $updateColumns = true,
         array $params = []
     ): static;
+
+    /**
+     * Returns copy of the instance with enabled or disabled typecasting of values when inserting or updating records.
+     */
+    public function withDbTypecasting(bool $dbTypecasting = true): static;
+
+    /**
+     * Returns a copy of the instance with enabled or disabled typecasting of values when retrieving records from DB.
+     */
+    public function withPhpTypecasting(bool $phpTypecasting = true): static;
+
+    /**
+     * Returns a copy of the instance with enabled or disabled typecasting of values when inserting, updating
+     * or retrieving records from DB.
+     */
+    public function withTypecasting(bool $typecasting = true): static;
 }