+++ src/QueryBuilder/AbstractDMLQueryBuilder.php
@@ -55,6 +55,8 @@ use function sort;
  */
 abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
 {
+    protected bool $typecasting = true;
+
     public function __construct(
         protected QueryBuilderInterface $queryBuilder,
         protected QuoterInterface $quoter,
@@ -145,6 +147,13 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         throw new NotSupportedException(__METHOD__ . ' is not supported by this DBMS.');
     }
 
+    public function withTypecasting(bool $typecasting = true): static
+    {
+        $new = clone $this;
+        $new->typecasting = $typecasting;
+        return $new;
+    }
+
     /**
      * Prepare traversable for batch insert.
      *
@@ -186,7 +195,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         /** @var string[] $names */
         $names = array_values($columnNames);
         $keys = array_fill_keys($names, false);
-        $columns = $this->schema->getTableSchema($table)?->getColumns() ?? [];
+        $columns = $this->typecasting ? $this->schema->getTableSchema($table)?->getColumns() ?? [] : [];
 
         foreach ($rows as $row) {
             $i = 0;
@@ -339,7 +348,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
         $names = [];
         $placeholders = [];
         $columns = $this->normalizeColumnNames($columns);
-        $tableColumns = $this->schema->getTableSchema($table)?->getColumns() ?? [];
+        $tableColumns = $this->typecasting ? $this->schema->getTableSchema($table)?->getColumns() ?? [] : [];
 
         foreach ($columns as $name => $value) {
             $names[] = $this->quoter->quoteColumnName($name);
@@ -373,7 +382,7 @@ abstract class AbstractDMLQueryBuilder implements DMLQueryBuilderInterface
     {
         $sets = [];
         $columns = $this->normalizeColumnNames($columns);
-        $tableColumns = $this->schema->getTableSchema($table)?->getColumns() ?? [];
+        $tableColumns = $this->typecasting ? $this->schema->getTableSchema($table)?->getColumns() ?? [] : [];
 
         foreach ($columns as $name => $value) {
             if (isset($tableColumns[$name])) {